Suitable for CodIgniter4 Thirdparty Library
You are supposed to use this library together with the the most appropriate Controller depending on your source code in php CodIgniter4. Make sure your database is functioning well if you are using mysqli and it is linked well with database it works well in Codigniter4 and backs up data in sql format.